Small changes for readability and efficiency
Hi Mike,

You've written largely readable code with good use of comments and mostly effective use of whitespace – I've made some small suggested edits for whitespace and readability in this pull request.

You have one section of code that is a good candidate for use of a loop to make some ellipses in a more efficient manner – we can look at this in class together.

On the whole, you've clearly put a good effort into this. Keep it up.

Take care,

Mr. Gordon